+++ back/CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 ### Added
-- customDisplayType in display_options"
+- customDisplayType in display_options
+- Route to get active courses by user
 
 ### Changed
 - OTHER connection type to TCP